Move SE2 background model to its own folder
