+++ SEImplementation/SEImplementation/Background/SE2/SE2BackgroundLevelAnalyzer.h
