+++ SEImplementation/SEImplementation/Background/SE2/SE2BackgroundUtils.h
@@ -26,7 +26,7 @@
 
 #include <sstream>
 #include "ElementsKernel/Logging.h"         // for Logging::LogMessageStream, etc
-#include "SEImplementation/Background/BackgroundDefine.h"
+#include "SEImplementation/Background/SE2/BackgroundDefine.h"
 
 namespace SourceXtractor {
