+++ SEImplementation/SEImplementation/Background/SE2/SplineModel.h
@@ -26,7 +26,7 @@
 #define	SPLINEMODEL_H
 
 #include <boost/filesystem.hpp>
-#include "SEImplementation/Background/BackgroundDefine.h"
+#include "SEImplementation/Background/SE2/BackgroundDefine.h"
 
 namespace SourceXtractor {
