+++ SEImplementation/SEImplementation/Background/SE2/TypedSplineModelWrapper.h
@@ -28,7 +28,7 @@
 #include <boost/filesystem.hpp>
 #include "SEFramework/Image/ImageBase.h"
 #include "SEFramework/Image/ImageSource.h"
-#include "SEImplementation/Background/SplineModel.h"
+#include "SEImplementation/Background/SE2/SplineModel.h"
 
 namespace SourceXtractor {
